requirements.txt
crewai==0.80.0
crewai-tools==0.17.0
arxiv==2.1.3
tavily-python==0.5.0
python-dotenv==1.0.1
anthropic==0.49.0
langchain-anthropic==0.3.10
